+++ src/types/result.ts
@@ -17,7 +17,7 @@ export interface RunResult {
 export interface Statement<T = unknown, P extends unknown[] = unknown[]> {
   all(...params: P): T[];
   get(...params: P): T | undefined;
-  run(...params: P): Promise<RunResult>;
+  run(...params: P): RunResult;
   finalize(): void;
 }